changed from dtype=numpy.int32 to dtype=int as suggested by original author

+++ kkcalc/data.py
@@ -377,7 +377,7 @@ def calculate_asf(Stoichiometry):
 		total_E = numpy.unique(total_E)
 		# add weighted asf data sets for KK calculation
 		total_Im_coeffs = numpy.zeros((len(total_E)-1, 5))
-		counters = numpy.zeros((len(Stoichiometry)),dtype=numpy.int32)
+		counters = numpy.zeros((len(Stoichiometry)),dtype=int)
 		for i,E in enumerate(total_E[1:]):
 			sum_Im_coeffs = 0
 			for j in range(len(counters)):
